pyproject.toml
[tool.poetry]
name = "lumiere"
version = "0.1.0"
description = ""
authors = ["SophiaBarroso <zastanisophia@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1"
djangorestframework = "^3.13.1"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
